fix buid plan hanging with --debug
When we look for os build plans, we could run into hanging the docker-machine
driver with using --debug.  This removes the debuging log output for
data returned on the call for get build plans.

We've also added an option to control debug output in our test cases
for the logger

Signed-off-by: Edward Raigosa <edward.raigosa@hpe.com>